# [test] [POC] Stop and start bitcoind nodes asynchronously in functional tests