@MarcoFalke 2017-03-22
I think it is good to keep python 3.4 as minimum version for now. Especially, since we plan to run the functional tests on `make check`.